Respect the order of existing resources when flattening workload components
This commit modifies the flattening logic of workload components to respect
the ordering of any existing components. Respecting ordering allows terraform
to produce a cleaner diff of when resources are changed.

As an example, if the following terraform configuration was defined:

```
container {
  name = "app"
  image = "my-container-image:latest"
}

container {
  name = "log-shipper"
  image = "my-log-shipper:latest"
}
```

and when our provider flattened the following API resposnse during a refresh:

```
container {
  name = "log-shipper"
  image = "my-log-shipper:latest"
}

container {
  name = "app"
  image = "my-container-image:latest"
}
```

This would cause terraform to believe the resource needs to be updated due to
the containers being reordered and should show a diff of each container being
removed and added in the correct order. Instead, we now flatten the definitions
with respect to their original order.