Patch API structs after they're generated
The network policy AH, ESP, GRE, and ICMP prootocols don't have any
parameters. If they're present in the network policy then they're on. As
such these should be empty structs instead of interfaces and pointers in
the upstream protocols struct. If the pointer is not nil then the
protocol is bound to the policy.