Use automatic token authentication
Automatic token authentication will allow the GitHub Action configuration
to define what permissions the implicit github token injected into the workflow
is given. This helps reduce the attack surface of providing an access token with
more than needed and reduces the number of github tokens we need to manage.

This adds the write permission for contents based on the issue opened on goreleaser
for what permissions are required.

References:
* https://github.com/goreleaser/goreleaser-action/issues/311
* https://docs.github.com/en/actions/security-guides/automatic-token-authentication