Switch to PUT
In "real life" this would need at least a refresh to get the version filled in.
That may be possible with a StateMigration function using a new
SchemaVersion for the resource. I tried to shim it by using PATCH
if we don't have the version, which will work for most changes.