Adjusted code for PUT handling
This is largely about sending complete values rather than
trying to keep account of prior state and sending semi-diffs.
As it is, I'm not entirely certain things were previously working as
expected. For example, when clearing probes, it looks like it would
have only sent the tiemout information, but wouldn't have sent
actually empty probes in the PATCH. That likely resulted in them
not in fact being removed.

I've added an additional test that verifies removal of probes
when they leave the config